feat(customer-head): add template string and json dynamic forms
