+++ src/pages/customer/dataschema.json
@@ -1,0 +1,50 @@
+{
+    "type": "object",
+    "properties": {
+        "jenisperumahan": {
+            "title": "Jenis Perumahan",
+            "type": "string",
+            "minLength": 1
+        },
+        "alamat": {
+            "title": "Alamat",
+            "type": "string"
+        },
+        "harga": {
+            "title": "Harga (Rp.)",
+            "type": "integer"
+        },
+        "description": {
+            "title": "Long Description",
+            "type": "string"
+        },
+        "done": {
+            "type": "boolean"
+        },
+        "lebar": {
+            "title": "Lebar (m)",
+            "type": "integer"
+        },
+        "panjang": {
+            "title": "Panjang (m)",
+            "type": "integer"
+        },
+        "arahmataangin": {
+            "title": "Arah Mata Angin",
+            "type": "string",
+            "enum": [
+                "Timur",
+                "Selatan",
+                "Barat",
+                "Utara"
+            ]
+        },
+        "tingkat": {
+            "title": "Tingkat",
+            "type": "integer"
+        }
+    },
+    "required": [
+        
+    ]
+}
